Reach green compile, e2e timedemo Tier 0, 9 test modules passing
Captures the working tree's green state after an in-progress module-
consolidation refactor (merging monster_exact/monster_object_interaction/
monster_dat/monster_new into monster.rs/monstdat.rs and reorganizing
toward the C++ source tree) plus the fixes needed to reach green.

Phase 0 - compile (3 errors -> 0):
- Recreate monster_exact.rs (adapter: re-export Monster/MonsterMode +
  MonsterManager) and monster_object_interaction.rs (door/blocking fns),
  adapted to the consolidated Monster API (x/y, hp/max_hp, position()).
- Add MonsterFlags::CAN_OPEN_DOOR + can_open_door().
- Fix API drift in game_state.rs / combat_integration.rs.
- Fix objdat.rs cross-context import (crate::levels -> super::types).
- Fix test/example call-site drift (Monster::new 4->5 args, GameState seed).

Phase 0b - baseline recorded in BASELINE.md (compile green; tests red
across modules + one concurrency-related stack-overrun crash, since
characterized as a data race over static state, not a single test).

Phase 1 - e2e timedemo Tier 0:
- engine/demo_reader.rs: real .dmo parser (header + event stream) ported
  from Source/engine/demomode.cpp, validated against the real
  WarriorLevel1to2 fixture.
- tests/timedemo_e2e.rs: Tier 0 green (parse + MPQ save detection);
  Tier 1..3 replay #[ignore]d pending engine systems.

Phase 2 - 9 modules green (~250 tests), 4 real C++-alignment bug fixes:
- monster (53), missiles (95): is_monster_mode_move mode match; missiles
  float arithmetic + screenToMissile formula.
- lighting (36): usize underflow in cone interpolation.
- combat_system (12), dialogue (32), gamemenu (10), capture (10),
  pathfinding (2): test/API drift + platform path fix.

Remaining red (diagnosed, needs focused work): inventory (grid encoding
design bug, 20+ sites), cel (transparent-run decoding), automap (iso
projection convention), player_controls (inventory slot layout).

+++ devilutionx-rs/src/game/lighting.rs
@@ -401,8 +401,8 @@ impl LightManager {
             for offset_x in 0..8 {
                 for y in 0..16 {
                     for x in 0..16 {
-                        let a = (8 * x - offset_x) as i32;
-                        let b = (8 * y - offset_y) as i32;
+                        let a = 8 * x as i32 - offset_x as i32;
+                        let b = 8 * y as i32 - offset_y as i32;
                         let distance = ((a * a + b * b) as f32).sqrt() as u8;
                         self.light_cone_interpolations[offset_x][offset_y][x][y] = distance;
                     }
